managed file: adds api to fetch a file
ManagedFile class now supports file download through
sync_from_resource() function.
Add a make_dist_folder() to create the target folder on the remote when
it does not exist to prevent failures from tools which require the
folder to exist.
rpath is now a property computed on the fly
File existence is now pushed back to sync_to_resource() because the
managed file may not exist (fetch case) when creating the instance.

Signed-off-by: Nicolas VINCENT <nicolas.vincent@vossloh.com>