client: fix console --loop on microcom error
In certain situation, e.g. if the serial port vanishes (i.e. on-board
FTDI chips on power off), microcom returns with 1 and:

  Got EOF from port
  connection lost

The loop should still continue, so do not raise an exception in this
case.

This uncovers another problem: if the corresponding place is released,
ser2net is stopped from the exporter and the port is set to None. Check
that after updating the resource to prevent a crude AssertionError later
due to an unavailable port.

Fixes: cdebb1cc ("remote/client: return exit code for ssh/scp/rsync/telnet/video/audio/console")
Signed-off-by: Bastian Krause <bst@pengutronix.de>