remote/exporter: Add option to default to fqdn for hostname
In distributed setups, the coordinator may point to exporters/places
in different subnets. In order for clients to resolve resources behind
the exporter, they need to know its FQDN instead of just the hostname.

Using an option to change the default avoids conflicts for configs which
rely on the hostname only.

Signed-off-by: Andreas Naumann <anaumann@ultratronik.de>