github/workflows: test Python 3.12
crossbar is not compatible with Python 3.12, yet: [1] and [2] are not
part of a release. A release will probably take some more time [3][4].

We shouldn't advertise Python 3.12 support until that happens. But we
can make sure that everything else keeps working fine with Python 3.12
by running the test suite. A previous commit marked the failing crossbar
tests as XFAIL for 3.12.

[1] https://github.com/crossbario/crossbar/pull/2091
[2] https://github.com/crossbario/crossbar/pull/2093
[3] https://github.com/crossbario/crossbar/pull/2093#issuecomment-1853007773
[4] https://github.com/crossbario/crossbar/pull/2091#issuecomment-1850126501

Signed-off-by: Bastian Krause <bst@pengutronix.de>