dockerfiles: Allow passing arbitrary arguments to build.sh
To allow customizing the arguments passed to docker/podman build(x)
beyond just the platform, modify build.sh to pass all its arguments
through to the build command.

For the platform case, this means it is now necessary to pass
"--platform linux/arm64" instead of just "linux/arm64" so modify
README.md to reflect this.

This change also has the effect of automatically using buildx if
supported, even if no arguments are passed, so move the information
about buildx further up in the text. The link to podman's buildx is
dropped as (unlike Docker) "podman buildx" is just an alias for
"podman build".

Signed-off-by: Peter Hoyes <Peter.Hoyes@arm.com>