pyproject.toml: prevent sphinxcontrib-devhelp incompatibilities
sphinxcontrib-devhelp>=1.0.3 depends on Sphinx>=5. This wasn't an issue
until now, since labgrid pins Sphinx==4.2.0 on stable. With
sphinxcontrib-devhelp>=1.0.6, this explicit dependency was dropped to
prevent cyclic dependencies, so the latest version gets installed,
although it's clearly incompatible. This is detected during runtime:

  The sphinxcontrib.devhelp extension used by this project needs at least Sphinx v5.0; it therefore cannot be built with this version.

So set sphinxcontrib-devhelp to a version lower than when the explicit
Sphinx dependency was dropped.

Signed-off-by: Bastian Krause <bst@pengutronix.de>