pyproject.toml: prevent sphinxcontrib-jsmath incompatibilities
sphinxcontrib-jsmath>=1.0.2 will depend on Sphinx>=5. This wasn't an issue
until now, since labgrid pins Sphinx==4.2.0 on stable. With
sphinxcontrib-jsmath>=2.0.2, this explicit dependency will be dropped to
prevent cyclic dependencies, so the latest version gets installed,
although it's clearly incompatible. This will be detected during runtime:

  The sphinxcontrib.jsmath extension used by this project needs at least Sphinx v5.0; it therefore cannot be built with this version.

So set sphinxcontrib-jsmath to a version lower than when the explicit
Sphinx dependency was dropped.

Signed-off-by: Bastian Krause <bst@pengutronix.de>