pyproject.toml: prevent sphinxcontrib-serializinghtml incompatibilities
sphinxcontrib-serializinghtml>=1.1.6 depends on Sphinx>=5. This wasn't an
issue until now, since labgrid pins Sphinx==4.2.0 on stable. With
sphinxcontrib-serializinghtml>=1.1.10, this explicit dependency was
dropped to prevent cyclic dependencies, so the latest version gets
installed, although it's clearly incompatible. This is detected during
runtime:

  The sphinxcontrib.serializinghtml extension used by this project needs at least Sphinx v5.0; it therefore cannot be built with this version.

So set sphinxcontrib-serializinghtml to a version lower than when the explicit
Sphinx dependency was dropped.

Signed-off-by: Bastian Krause <bst@pengutronix.de>