crossbar: enable auto fragmentation of outgoing WebSocket messages
Larger labgrid remote infrastructure setups with a big number of
exported resources are limited by the maximum WebSocket payload size of
1M in autobahn [1].

Future patches in labgrid will increase the maximum payload size from 1M
to 10M to allow larger setups.

In oder to be able to send these payloads, enable auto fragmentation of
outgoing WebSocket messages into multiple WebSocket frames [2] in the
crossbar configuration. Use 65536 bytes which is the value set in the
ApplicationRunner [3].

[1] https://github.com/crossbario/autobahn-python/blob/359f868f9db410586cf01c071220994d8d7f165a/autobahn/asyncio/wamp.py#L223
[2] https://github.com/crossbario/crossbar/blob/master/docs/WebSocket-Options.rst
[3] https://github.com/crossbario/autobahn-python/blob/359f868f9db410586cf01c071220994d8d7f165a/autobahn/asyncio/wamp.py#L224

Signed-off-by: Bastian Krause <bst@pengutronix.de>