doc/configuration: update/fix driver bindings
Initially, all driver bindings were documented with their respective
binding key. This wasn't documented consistently since then.

Especially when using multiple instances of a single driver class, the
bindings are needed to specify on which resource or other driver the
driver should bind on. Until now, most binding keys had to be looked up
in the source code. So improve this situation by fixing and adding
binding keys as well as also adding all Remote/Network resource variants
a driver is able to bind to.

Signed-off-by: Bastian Krause <bst@pengutronix.de>