resource: udev: also suggest ID_USB_INTERFACE_NUM for USBResource
A USB serial device often provides multiple interfaces for the same
serial number. In these cases it is not possible to distinguish
the interfaces solely based on the serial number (ID_SERIAL_SHORT).
However, if additionally the interface number (ID_USB_INTERFACE_NUM)
is provided, it is possible to distinguish all interfaces without
the need of using the full device path (ID_PATH).

Example:

  (...)
  === suggested matches ===
  (...)
  USBSerialPort:
    match:
      ID_SERIAL_SHORT: ABCDEF00001
      ID_USB_INTERFACE_NUM: '00'
  (...)
  === suggested matches ===
  (...)
  USBSerialPort:
    match:
      ID_SERIAL_SHORT: ABCDEF00001
      ID_USB_INTERFACE_NUM: '01'

Signed-off-by: Joerg Hofrichter <joerg.hofrichter@ni.com>