resource: matchedsysfsgpio: add udev matched gpios
Add a support gpio line for which the gpiochip is identified
by udev and pin number provided is local the identified gpiochip.

This is useful when there is several gpio controller and the order
in which the controller are probed is not predictable. An common
example are the GPIOs provided by some USB FTDI chips.

This resouce re-use most the support already done for SysfsGPIO.
The global gpio index is computed at runtime, based on the udev
properties of the matched gpiochip.

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>