driver: SigrokDriver capture file is transferred to wrong path
when specified filename is not absolute and using a remote sigrok resource

Usually when an absolute path is supplied as `filename`
the call `abs.path.join()` resolves to `filename`.
But when the file path is relative,
it resolves it to a relative path appended to the tmp dir,
instead of from where the user calls this method.

Signed-off-by: Felix Zwettler <felix.zwettler@duagon.com>