doc/configuration: UUUDriver: align example with reality
Besides the image name, the UUUDriver supports a script attribute.
This is described in the Arguments: section, but not in the example,
which uses the non-existent cmd attribute instead.

Fix that and while at it, correctly indicate that script is optional.
Omitting it will just remove the -b $script from the uuu invocation.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>