shelldriver: allow dashes and dots in device names
The `get_ip_addresses` may be called with a specific device and returns
alls IP addresses on that interface. Due to a limitation in the regex,
devices including either dashes or dots aren't detected.

On the OpenWrt distributions the default device is called `br-lan`,
usually connecting ethernet ports and WiFi devices. Likewise dots are
attached to the device name to specify VLANs.

This commit extends the regex to support both dashes and dots.

Signed-off-by: Paul Spooren <mail@aparcar.org>