exporter: check ser2net version in exporter
The following ser2net versions return nonzero even on a successful use
of the -v flag to get the version:

v4.4.0
v4.5.0
v4.5.1
v4.6.0
v4.6.1

This was fixed in v4.6.2, which was only recently released. Modify
_start() so that it uses subprocess.run() to capture stdout and the
returncode, then check against the version and effectively ignore the
return value if there's a match. If the ser2net version is one of the
unaffected ones (e.g. v4.6.2) and still returns nonzero, then raise an
ExporterError.

Signed-off-by: Trevor Gamblin <tgamblin@baylibre.com>