driver: Provide a driver for writing U-Boot to a board
Each SoC has its own way of booting and this means that the firmware
layout for each SoC is typically different. On top of that, some SoCs
allow writing an image over USB. which uses a different format, e.g.
separate binaries for each stage.

Trying to make all this work is extremely challenging. Labgrid provides
a nice unification point, where the method appropriate for each SoC can
be included, with a small amount of configuration to control things.

Binary blobs can be provided to make the writer work correctly.

This driver is capable of writing SD-card images for several popular
devices as well as handling USB booting in some cases. It also supports
QEMU. It can be easily extended as necessary.

Signed-off-by: Simon Glass <sjg@chromium.org>