driver/usbstoragedriver: Adjust how sync works
The current approach is slow, since each block must be individually
written to the disk before the next can start.

The more common approach is to do a manual 'sync' after writing. Adjust
write_image() to use this approach. It is several times faster on
non-trivial files, even with the need for the sync.

Signed-off-by: Simon Glass <sjg@chromium.org>