protocol: Allow more control of reset
The reset protocol only supports doing a reset. Sometimes we want to
assert reset while something is done (e.g. a power cycle or a recovery
signal), then de-assert it afterwards.

Add a new set_reset_enable() method for this.

The name is a little strange: one might expect set_enable() instead. But
with things like the PowerResetMixin we want to avoid ambiguity over
whether this refers to powering the device or resetting it.

Signed-off-by: Simon Glass <sjg@chromium.org>