strategy: Show the U-Boot version when booting
At present the 'start' state of U-Boot consumes all the initial U-Boot
output, including the SPL and proper headers and any version
information. This makes it impossible to run the U-Boot pytests, since
this output is expected from the board.

There are two ways to fix this:
- Show the output
- Print a message indicating that U-Boot has booted, so tests don't
  wait for the headers

Do both, the first to make tests work existing U-Boot work with tests,
the second to resolve the problem for future U-Boot versions, once the
U-Boot pytest patch is accepted:

   test: Allow signaling that U-Boot is ready

Series-changes: 8
- Provide a flag to indicate whether the output should be cleared

Signed-off-by: Simon Glass <sjg@chromium.org>