labgrid: Add a way to query information from a driver
When integrating Labgrid into other projects it is useful to be able to
query information known by a particular driver. For example, U-Boot
wants to know the board name of a particular target, so that it can
pass this to pytest.

Add a new driver method to provide this feature.

Signed-off-by: Simon Glass <sjg@chromium.org>