labgrid: Add a way to query information from a driver
When integrating Labgrid into other projects it is useful to be able to
query information known by a particular driver. For example, U-Boot
wants to know the board name of a particular target, so that it can
pass this to pytest.

Add a new driver method to provide this feature.

Series-changes: 6
- Show a more friendly error when the wrong query arguments are given

Series-changes: 7
- Collect all info before printing anything
- Make the query synchronous
- Drop the extra argument from print_version()

Series-changes: 8
- Show an empty line when the value is None

Signed-off-by: Simon Glass <sjg@chromium.org>