remote/client: Support finding a place by role
The role (or is it target?) is used in the environment and allows
multiple roles to use the same place. It is therefore often better to
specify the role to use rather than the place, since the place can be
determined from the role.

Add a new -r option to support this.

Signed-off-by: Simon Glass <sjg@chromium.org>