strategy: Support boards which need a recovery button
Some boards boot from internal flash unless a 'recovery' button is
pressed. Add support for this, so that these boards can be used in a
lab.

When Labgrid fails to get to the U-Boot command line, show the output
that was received, so that the user can see what is going on.

Series-changes: 7
- Allow recovery to be missing for the board

Signed-off-by: Simon Glass <sjg@chromium.org>