remote/client: Provide an internal console
At present Labgrid uses microcom as its console. This has some
limitations:

- console output is lost between when the board is reset and microcom
  connects
- txdelay cannot be handled in microcom, meaning that boards may fail
  to receive expected output
- the console may echo a few characters back to the caller in the time
  between when 'labgrid-client console' is executed and when microcom
  starts (which causes failures with U-Boot test system)

For many use cases, microcom is more than is needed, so provide a simple
internal terminal which resolved the above problems.

It is enabled by a '-i' option to the 'console' command, as well as an
environment variable, so that it can be adjusted without updating a lot
of scripts.

When the console is connected, a message appears:

   Terminal ready...press Ctrl-] twice to exit

This message is automatically cleared when console output is received
from the board, to provide a seamless interactive experience.

The logfile is supported, as with the external console.

To exit, press Ctrl-] twice, quickly.

Series-version: 5
Series-changes: 5
- Drop an incorrect mention of microcom in the internal-console comment
- Update the commit message to mention the message on connection

Series-changes: 4
- Get internal console working with qemu
- Show a prompt when starting, to indicate it is waiting for the board

Signed-off-by: Simon Glass <sjg@chromium.org>