util/agent: disable pylint warnings
The pylint code warns because of a possibly-used-before-assignment,
which is only valid in case python gobject introspectiuon is not used.
The reason the variable is hidden behind a type check is a hang related
to sphinx, which when trying to import the module for autodoc hangs
forever. The normal case is that the `nm` variable is always assigned
before use, so disable the related warnings.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>