util/yaml: warn on duplicate keys in mappings
When adding drivers/resources to an environment config or resources to
an exporter config, users can stumble upon redundant keys. The resulting
errors are not obvious, the last key simply overrides any previous
entries with the same key.

In order to make these config errors more obvious, emit a warning when
duplicate keys are detected in mappings.

Signed-off-by: Bastian Krause <bst@pengutronix.de>