remote/config: simplify ResourceConfig template loading
ResourceConfig passes a template loader with the directory name of the
resource (exporter) config to jinja2.Environment(). Instead of using
the Environment's loader to load the actual config, the config file is
opened and read from manually and its content is then passed to
jinja2.Environment.from_string().

Change that by using the designated jinja2.Environment.get_template()
method to load the config.

Signed-off-by: Bastian Krause <bst@pengutronix.de>