protocol/driver: add `DigitalInputProtocol`, `GpioDigitalInputDriver`
this adds a new protocol `DigitalInputProtocol` and a GPIO driver
that implements it.
The `DigitalOutputProtocol` ABC now also inherits it, ensuring it can be
used in all existing driver implementations.
Additionally the client now automatically uses `DigitalInputProtocol`
when the command is `io get`.

By seperating the input from output protocol it is possible
to make sure the GPIO line is configured in direction `input` in sysfs,
which avoids possible shorts when reading out digital states from DUT's.
Otherwise, one would have to rely on the circuitry of the GPIO line to
prevent that.

Signed-off-by: Felix Zwettler <Felix.Zwettler@duagon.com>