remote: use more explicit event loop handling
Calling asyncio.get_event_loop() with no current event loop is deprecated
since Python 3.10 and will be an error in some future Python release
[1].

Whenever we don't expect to run with an event loop, create one
explicitly. In coroutine and callbacks from asynchronous code, use the
more explicit asyncio.get_running_loop() to get the loop.

Note that this does not work in
labgrid.resources.ethernetport.EthernetPortManager: This code is usually
not called in coroutines and callbacks from asynchronous code, so
asyncio.get_running_loop() does not work there. So stick to
asyncio.get_event_loop() there and just expect to be called with a
running event loop (which is the non-deprecated use case for this
function).
Users that do not have an event loop running will see a justified
DeprecationWarning with Python >= 3.12 and an error in some future Python
version.

[1] https://docs.python.org/3/library/asyncio-eventloop.html#asyncio.get_event_loop

Signed-off-by: Bastian Krause <bst@pengutronix.de>