fix: match interface regex in ShellDriver
Fix regex to correctly match network interfaces with dots and/or dashes in the name

Previously, the regex for extracting the interface from the default route was not capable of matching network interfaces containing dots ('.') and/or dashes ('-') in their names. This commit improves the regex by allowing such interfaces to be correctly identified.

For example, interfaces like 'br-lan.42' and 'eth0-1' are now correctly matched by the updated regex.

Sample output of the `ip -4 route list default` command:
default via 192.168.42.255 dev br-lan.42  src 192.168.42.1

Signed-off-by: Rainer Poisel <rainer@honeytreelabs.com>