driver/usbvideodriver: reduce latency of GStreamer pipeline
The playbin3 uses a urisourcebin to open and read the uri. The
urisourcebin detects the fd as a stream uri and internally crates a
multiqueue with a default buffer duration of 5 s and a high watermark of
0.60 to buffer the stream. This adds 3 s latency to the video stream.

If the source signals that is it a live source, the urisourcebin would
avoid buffering of the stream. Unfortunately, the fdsrc doesn't signal
that the data is live, and doesn't have a property to change this.

Use buffer-duration to set the size of the queue for buffering to 0 to
avoid this additional latency.

The queue could be completely removed by setting "buffering" in the
flags property of playbin3 to 0, which would be equal to the behavior of
a live source, but the changed flags are more difficult to understand
than the buffer duration.