resource: Add generic USB port
This is useful, if the DUT is a USB device, where you are not yet sure,
what exactly it is. This allows the test to use the wrapper to
communicate with the device, even if it is on the network.

Might also be useful, if you don't yet have a driver for a USB
controlled device and want to test it out before upstreaming it.

Signed-off-by: Fabian Pflug <fabian.pflug@grandcentrix.net>