target: store features and options
Instead of always having to revert to the configuration for options and
features, store both directly in the target object. Also deprecate the
setters and getters for both in the configuration object.

Reasoning is that setters are not required for this, the YAML
configuration should be changed instead. The getters are not required if
the options and features are bound to the target which is more intuitive
anyway.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>