driver: digitalloggers_restapi: enable REST API
The legacy HTTP API does not work on newer units without manually
changing the configuration in Setup.

This driver is based on https://www.digital-loggers.com/restapi.pdf

Curl examples that were used for development are included as comments
so that users can test their access outside of labgrid.

The REST API seems to only allow authenticated users, so the host:
parameter is parsed to pass user and password to HTTPDigestAuth.
CSRF is also required, so a valid (simple) header is provided.
Non-authenticated URLs are supported, but most likely will not
work.

HTTPS is recommended, but the units ship with self-signed certificates
so SSL certificate verification warnings are intentionally ignored.

Add to test_powerdriver.py import test.

Example usage in lg-env.yaml (default as-shipped settings):

NetworkPowerPort:
    model: 'digitalloggers_restapi'
    host: 'http://admin:1234@192.168.0.100'
    index: 0

Signed-off-by: Tim Orling <tim.orling@konsulko.com>