util/agents/usb_hid_relay: fix concurrent access
Since c9fc5bfdb9da, it was no longer possible to use the same USB relay
device from multiple labgrid processes, as the USB device was kept open
and claimed.

To fix this, we use a context manager which first claims the USB
interface (with retry while busy) and releases it after the transaction.

With this fix, multiple processes can toggle outputs in a busy loop
without causing 'USBError(16, 'Resource busy')' failures.

Fixes: c9fc5bfdb9da ("labgrid/util/agents/usb_hid_relay: keep the USB device open")
Signed-off-by: Jan Luebbe <jlu@pengutronix.de>