driver/power/gude8031: Add Support for Gude 87-1210-18
The Gude Expert Power Control 87-1210-18 is a vertical 20 port power
distribution unit:
https://gude-systems.com/en/products/expert-power-control-87-1210/

It uses the same HTTP-API, as the Gude Power Control 8031 - but simply
has more ports.
This commit adds support and documentation for this device.

Loosing the restrictions on the `index` makes it possible to re-use this
driver.

Setting a non-existent port fails silently.
But getting a non-existent port will cause an exception.
I deem this behavior acceptable, since a new device will be tested
during hardware setup anyway.
So we do not need to check the actual number of ports during normal
operation.

Signed-off-by: Chris Fiege <cfi@pengutronix.de>