remote: add support for pystuck
To allow debugging in production, we need a way to introspect objects in
the coordinator and exporter without stopping the process. This is
possible with pystuck.

We add it as a dev dependency only, as it's only needed in special
circumstances and can be installed manually in that case. It is only
enabled when the --pystuck option is passed on startup.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>