add USBHub resource and export
This resource is a bit unusual, as there is no corresponding driver, as
it is only useful to monitor whether the expected USB hubs are detected
by an exporter.

In larger labs, there are often USB hubs that are considered part of the
infrastructure. When these fail (either temporarily or permanently), the
connected devices aren't detected any more. For users of individual
places, this leads to unavailable resources without a clear cause.

By configuring USBHub resources for the infrastructure hubs in the
exporters, we can start monitoring hub availability centrally via the
coordinator.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>