pytestplugin: make "target" a parametrized fixture
This allows users to run their pytest test scripts with an
LG_ENV/--lg-env that has multiple targets instead of forcing the user to
define one named "main" in order to run the tests.

One of the downsides of this implementation is that autogenerated doc
wouldn't define the "target" fixture anymore as it is entirely
programmatically defined.

Signed-off-by: Quentin Schulz <quentin.schulz@cherry.de>