qemudriver: add a qemu-default display option
The barebox test suite adds an --interactive option to pytest, which
allows starting QEMU interactively with the Labgrid Environment YAML to
allow manual testing.

So far, it wasn't possible to use this with an interactive graphic output,
because Labgrid allowed only headless displays or none.

none is the default and will configure QEMU with -nographic, which
can not be overridden later on and -display also has no default value.

To support this use case, add a qemu-default option, which instructs
Labgrid to not touch -nographic and -display at all.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>