remote/coordinator: fail acquire for orphaned
Check if a place acquire request tries to lock a resource that is marked
as orphaned. This means we no longer need to reacquire orphaned resource
before trying to acquire a places resources since we now refuse to
acquire the resources.

This avoids long delays on aquire calls if the exporter responsible for
an (unrelated) orphaned resource doesn't process commands quickly.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>