util/managedfile: add cleanup_resource() method
When using ManagedFile with remote resources without NFS, files copied
to the remote by sync_to_resource() method will not be deleted
automatically.

While it is of-course possible to use something like systemd-tmpfiles
to cleanup after a while, files might pile up and cause out of diskspace
situations before systemd-tmpfiles gets to it.

This method can be used by users of ManagedFile objects to cleanup the
remote copy of the file when it is no longer needed.

Signed-off-by: Esben Haabendal <esben@geanix.com>